+++ jopa-owl2java/src/main/java/cz/cvut/kbss/jopa/owl2java/Constants.java
@@ -34,7 +34,7 @@ public class Constants {
     /**
      * Tool version.
      */
-    public static final String VERSION = "0.8.4";
+    public static final String VERSION = "$VERSION$";
 
 
     private Constants() {