+++ ontodriver-owlapi/src/main/java/cz/cvut/kbss/ontodriver/owlapi/connector/Connector.java
@@ -1,11 +1,11 @@
 /**
  * Copyright (C) 2016 Czech Technical University in Prague
- *
+ * <p>
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation, either version 3 of the License, or (at your option) any
  * later version.
- *
+ * <p>
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
@@ -16,6 +16,7 @@ package cz.cvut.kbss.ontodriver.owlapi.connector;
 
 import org.semanticweb.owlapi.model.OWLOntologyChange;
 
+import java.net.URI;
 import java.util.List;
 import java.util.function.Consumer;
 import java.util.function.Function;
@@ -28,6 +29,13 @@ import java.util.function.Function;
 public interface Connector {
 
     /**
+     * Gets logical URI of the ontology loaded by this connector.
+     *
+     * @return Ontology URI
+     */
+    URI getOntologyUri();
+
+    /**
      * Gets snapshot of the underlying ontology.
      * <p>
      * The snapshot is completely independent of the live ontology, so any changes to either are not visible to the
@@ -37,8 +45,20 @@ public interface Connector {
      */
     OntologySnapshot getOntologySnapshot();
 
+    /**
+     * Executes read-only operation on the live ontology.
+     *
+     * @param function The function to execute
+     * @param <R>      Result type
+     * @return Read result
+     */
     <R> R executeRead(Function<OntologySnapshot, R> function);
 
+    /**
+     * Executes a write operation on the live ontology.
+     *
+     * @param function The function to execute
+     */
     void executeWrite(Consumer<OntologySnapshot> function);
 
     /**
@@ -50,4 +70,13 @@ public interface Connector {
      * @param changes The changes to apply
      */
     void applyChanges(List<OWLOntologyChange> changes);
+
+    /**
+     * Closes the specified transactional snapshot.
+     * <p>
+     * This in particular means destroying the transactional ontology.
+     *
+     * @param snapshot The snapshot to close
+     */
+    void closeSnapshot(OntologySnapshot snapshot);
 }