+++ ontodriver-owlapi/src/test/java/cz/cvut/kbss/ontodriver/owlapi/connector/BasicStorageConnectorTest.java
@@ -1,11 +1,11 @@
 /**
  * Copyright (C) 2016 Czech Technical University in Prague
- *
+ * <p>
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation, either version 3 of the License, or (at your option) any
  * later version.
- *
+ * <p>
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
@@ -17,6 +17,7 @@ package cz.cvut.kbss.ontodriver.owlapi.connector;
 import cz.cvut.kbss.ontodriver.OntologyStorageProperties;
 import cz.cvut.kbss.ontodriver.config.Configuration;
 import cz.cvut.kbss.ontodriver.owlapi.OwlapiDataSource;
+import cz.cvut.kbss.ontodriver.owlapi.environment.Generator;
 import cz.cvut.kbss.ontodriver.owlapi.exception.InvalidOntologyIriException;
 import cz.cvut.kbss.ontodriver.owlapi.util.MutableAddAxiom;
 import org.junit.After;
@@ -28,6 +29,7 @@ import java.io.File;
 import java.net.URI;
 import java.nio.file.Files;
 import java.util.Collections;
+import java.util.Set;
 
 import static org.junit.Assert.*;
 
@@ -35,7 +37,10 @@ public class BasicStorageConnectorTest {
 
     private static final URI ONTOLOGY_URI = URI.create("http://krizik.felk.cvut.cz/ontologies/jopa/connector");
 
-    private AbstractConnector connector;
+    private BasicStorageConnector connector;
+
+    private OWLOntologyManager manager;
+    private OWLOntology ontology;
 
     @After
     public void tearDown() throws Exception {
@@ -51,29 +56,31 @@ public class BasicStorageConnectorTest {
 
     @Test
     public void loadsExistingOntology() throws Exception {
-        final URI physicalUri = initOntology();
+        final URI physicalUri = initOntology(Collections.emptySet());
         this.connector = new BasicStorageConnector(new Configuration(initStorageProperties(physicalUri, null)));
         assertNotNull(connector);
         assertTrue(connector.isOpen());
         final OntologySnapshot snapshot = connector.getOntologySnapshot();
         assertNotNull(snapshot.getOntology());
-        assertEquals(IRI.create(ONTOLOGY_URI), snapshot.getOntology().getOntologyID().getOntologyIRI().get());
         assertNotNull(snapshot.getOntologyManager());
         assertNotNull(snapshot.getDataFactory());
     }
 
-    private URI initOntology() throws Exception {
+    private URI initOntology(Set<OWLAxiom> axioms) throws Exception {
         final File targetFile = Files.createTempFile("connectortest", ".owl").toFile();
         targetFile.deleteOnExit();
         final OWLOntologyManager om = OWLManager.createOWLOntologyManager();
         final OWLOntology o = om.createOntology(IRI.create(ONTOLOGY_URI));
+        om.addAxioms(o, axioms);
         om.saveOntology(o, IRI.create(targetFile));
+        this.manager = om;
+        this.ontology = o;
         return targetFile.toURI();
     }
 
     @Test(expected = InvalidOntologyIriException.class)
     public void throwsExceptionWhenLoadedOntologyHasDifferentIri() throws Exception {
-        final URI physicalUri = initOntology();
+        final URI physicalUri = initOntology(Collections.emptySet());
         final URI logicalUri = URI.create("http://krizik.felk.cvut.cz/ontologies/jopa/different");
         this.connector = new BasicStorageConnector(new Configuration(initStorageProperties(physicalUri, logicalUri)));
     }
@@ -92,7 +99,7 @@ public class BasicStorageConnectorTest {
 
     @Test
     public void getSnapshotReturnsDistinctSnapshots() throws Exception {
-        final URI physicalUri = initOntology();
+        final URI physicalUri = initOntology(Collections.emptySet());
         this.connector = new BasicStorageConnector(new Configuration(initStorageProperties(physicalUri, null)));
         final OntologySnapshot snapshotOne = connector.getOntologySnapshot();
         final OntologySnapshot snapshotTwo = connector.getOntologySnapshot();
@@ -102,7 +109,7 @@ public class BasicStorageConnectorTest {
 
     @Test(expected = IllegalStateException.class)
     public void throwsExceptionWhenTryingToGetSnapshotOfClosedConnector() throws Exception {
-        final URI physicalUri = initOntology();
+        final URI physicalUri = initOntology(Collections.emptySet());
         this.connector = new BasicStorageConnector(new Configuration(initStorageProperties(physicalUri, null)));
         connector.close();
         assertFalse(connector.isOpen());
@@ -111,7 +118,7 @@ public class BasicStorageConnectorTest {
 
     @Test(expected = IllegalStateException.class)
     public void throwsExceptionWhenApplyChangesCalledOnClose() throws Exception {
-        final URI physicalUri = initOntology();
+        final URI physicalUri = initOntology(Collections.emptySet());
         this.connector = new BasicStorageConnector(new Configuration(initStorageProperties(physicalUri, null)));
         connector.close();
         assertFalse(connector.isOpen());
@@ -120,7 +127,7 @@ public class BasicStorageConnectorTest {
 
     @Test
     public void applyChangesModifiesTheCentralOntology() throws Exception {
-        final URI physicalUri = initOntology();
+        final URI physicalUri = initOntology(Collections.emptySet());
         this.connector = new BasicStorageConnector(new Configuration(initStorageProperties(physicalUri, null)));
         final OntologySnapshot snapshot = connector.getOntologySnapshot();
         final OWLClass cls = addClassToOntology(snapshot);
@@ -140,7 +147,7 @@ public class BasicStorageConnectorTest {
 
     @Test
     public void successfullySavesOntologyOnClose() throws Exception {
-        final URI physicalUri = initOntology();
+        final URI physicalUri = initOntology(Collections.emptySet());
         final OntologyStorageProperties storageProperties = initStorageProperties(physicalUri, null);
         this.connector = new BasicStorageConnector(new Configuration(storageProperties));
         final OWLClass cls = addClassToOntology(connector.getOntologySnapshot());
@@ -149,5 +156,55 @@ public class BasicStorageConnectorTest {
         this.connector = new BasicStorageConnector(new Configuration(storageProperties));
         final OntologySnapshot res = connector.getOntologySnapshot();
         assertTrue(res.getOntology().containsClassInSignature(cls.getIRI()));
+    }
+
+    @Test
+    public void getSnapshotCreatesNewAnonymousTransactionalOntology() throws Exception {
+        final URI physicalUri = initOntology(Collections.emptySet());
+        final OntologyStorageProperties storageProperties = initStorageProperties(physicalUri, null);
+        this.connector = new BasicStorageConnector(new Configuration(storageProperties));
+        final OntologySnapshot snapshot = connector.getOntologySnapshot();
+        assertTrue(snapshot.getOntology().getOntologyID().isAnonymous());
+    }
+
+    @Test
+    public void getSnapshotCopiesAxiomsIntoTheTransactionOntology() throws Exception {
+        final Set<OWLAxiom> axioms = Generator.generateAxioms();
+        final URI physicalUri = initOntology(axioms);
+        final OntologyStorageProperties storageProperties = initStorageProperties(physicalUri, null);
+        this.connector = new BasicStorageConnector(new Configuration(storageProperties));
+        final OntologySnapshot snapshot = connector.getOntologySnapshot();
+        final Set<OWLAxiom> transactionalAxioms = snapshot.getOntology().getAxioms();
+        assertTrue(transactionalAxioms.containsAll(axioms));
+    }
+
+    @Test
+    public void closeSnapshotRemovesTransactionalOntologyFromManager() throws Exception {
+        final URI physicalUri = initOntology(Collections.emptySet());
+        final OntologyStorageProperties storageProperties = initStorageProperties(physicalUri, null);
+        this.connector = new BasicStorageConnector(new Configuration(storageProperties));
+        final OntologySnapshot snapshot = connector.getOntologySnapshot();
+        final OWLOntology transactionalOntology = snapshot.getOntology();
+        final OWLOntologyManager manager = snapshot.getOntologyManager(); // We know this is the root manager
+        assertTrue(manager.contains(transactionalOntology));
+        connector.closeSnapshot(snapshot);
+        assertFalse(manager.contains(transactionalOntology));
+    }
+
+    @Test
+    public void getSnapshotResolvesImportsOfTheTransactionalSnapshot() throws Exception {
+        final URI physicalUri = initOntology(Collections.emptySet());
+        final String importedOnto = "http://onto.fel.cvut.cz/ontologies/ufo";
+        final OWLImportsDeclaration importDecl = manager.getOWLDataFactory()
+                                                        .getOWLImportsDeclaration(IRI.create(importedOnto));
+        manager.applyChange(new AddImport(ontology, importDecl));
+        manager.saveOntology(ontology, IRI.create(physicalUri));
+        final OntologyStorageProperties storageProperties = initStorageProperties(physicalUri, null);
+        this.connector = new BasicStorageConnector(new Configuration(storageProperties));
+        final OntologySnapshot snapshot = connector.getOntologySnapshot();
+        final Set<OWLOntology> imports = snapshot.getOntology().getImports();
+        assertEquals(1, imports.size());
+        final OWLOntology imported = imports.iterator().next();
+        assertEquals(IRI.create(importedOnto), imported.getOntologyID().getOntologyIRI().get());
     }
 }