+++ ontodriver-owlapi/src/test/java/cz/cvut/kbss/ontodriver/owlapi/environment/TestUtils.java
@@ -1,11 +1,11 @@
 /**
  * Copyright (C) 2016 Czech Technical University in Prague
- *
+ * <p>
  * This program is free software: you can redistribute it and/or modify it under
  * the terms of the GNU General Public License as published by the Free Software
  * Foundation, either version 3 of the License, or (at your option) any
  * later version.
- *
+ * <p>
  * This program is distributed in the hope that it will be useful, but WITHOUT
  * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
  * FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more
@@ -21,15 +21,11 @@ import org.semanticweb.owlapi.model.OWLOntology;
 import org.semanticweb.owlapi.model.OWLOntologyManager;
 import org.semanticweb.owlapi.reasoner.OWLReasoner;
 
-import java.util.Random;
-
 public class TestUtils {
 
     private TestUtils() {
         throw new AssertionError();
     }
-
-    private static final Random RANDOM = new Random();
 
     /**
      * Initializes a real ontology, which can be used in tests. It is not saved anywhere, so once the manager is
@@ -43,9 +39,5 @@ public class TestUtils {
         final OWLOntology ontology = manager
                 .createOntology(IRI.create("http://krizik.felk.cvut.cz/ontologies/adapterTest"));
         return new OntologySnapshot(ontology, manager, manager.getOWLDataFactory(), reasoner);
-    }
-
-    public static int randomInt(int max) {
-        return RANDOM.nextInt(max);
     }
 }